chore(types-styling): change type definition location and fix styling bugs
